% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloOAuth.R
\name{armadillo.get_token}
\alias{armadillo.get_token}
\title{Get ID Token}
\usage{
armadillo.get_token(server)
}
\arguments{
\item{server}{the URL of the Armadillo server}
}
\value{
The ID token string
}
\description{
Get an ID token to log in on an Armadillo server
}
