% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloConnection.R
\docType{class}
\name{ArmadilloConnection-class}
\alias{ArmadilloConnection-class}
\title{Class ArmadilloConnection.}
\description{
An Armadillo connection implementing the DataSHIELD Interface (DSI)
\code{\link[DSI]{DSConnection-class}}.
}
\section{Slots}{

\describe{
\item{\code{name}}{The name of the connection}

\item{\code{handle}}{The handle used to connect with the server}

\item{\code{user}}{The username used to authenticate}

\item{\code{cookies}}{The cookies set by the server}

\item{\code{token}}{An optional authentication token (JWT)}
}}

\keyword{internal}
