% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloOAuth.R
\docType{class}
\name{ArmadilloCredentials-class}
\alias{ArmadilloCredentials-class}
\title{ArmadilloCredentials Class}
\description{
An S4 class to represent authentication credentials used for accessing Armadillo servers.
}
\section{Slots}{

\describe{
\item{\code{access_token}}{Character. The access token used for authentication.}

\item{\code{expires_in}}{Numeric. The number of seconds until the token expires.}

\item{\code{expires_at}}{POSIXct. The timestamp when the token expires.}

\item{\code{id_token}}{Character. The ID token containing identity information.}

\item{\code{refresh_token}}{Character. The token used to obtain a new access token when expired.}

\item{\code{token_type}}{Character. The type of token (typically "Bearer").}

\item{\code{auth_type}}{Character. the authentication provider type (e.g., "keycloak", "fusionauth").}
}}

\keyword{internal}
