% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloDriver.R
\docType{class}
\name{ArmadilloDriver-class}
\alias{ArmadilloDriver-class}
\title{Class ArmadilloDriver with constructor armadillo}
\description{
An Armadillo DataSHIELD Service Driver implementing the DataSHIELD Interface
(DSI) \code{\link[DSI]{DSDriver-class}}. This class should always be initialized
with the \code{\link{armadillo}} function.
It returns a singleton that allows you to connect to Armadillo
}
\keyword{internal}
