% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloResult.R
\docType{class}
\name{ArmadilloResult-class}
\alias{ArmadilloResult-class}
\title{Class ArmadilloResult.}
\description{
An Armadillo result implementing the DataSHIELD Interface (DSI)
\code{\link[DSI]{DSResult-class}}.
}
\section{Slots}{

\describe{
\item{\code{conn}}{The connection used to create this result}

\item{\code{rval}}{The result}
}}

\keyword{internal}
