% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloConnection.R
\name{dsAggregate,ArmadilloConnection-method}
\alias{dsAggregate,ArmadilloConnection-method}
\title{Aggregate data}
\usage{
\S4method{dsAggregate}{ArmadilloConnection}(conn, expr, async = TRUE)
}
\arguments{
\item{conn}{\code{\link{ArmadilloConnection-class}} object.}

\item{expr}{Expression to evaluate.}

\item{async}{Whether the result of the call should be retrieved
asynchronously. When TRUE (default) the calls are parallelized over the
connections, when the connection supports that feature, with an extra
overhead of requests.}
}
\description{
Aggregate some data from the DataSHIELD R session using a valid R expression.
The aggregation expression must satisfy the data repository's DataSHIELD
configuration.
}
