% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloConnection.R
\name{dsAssignExpr,ArmadilloConnection-method}
\alias{dsAssignExpr,ArmadilloConnection-method}
\title{Assign the result of an expression}
\usage{
\S4method{dsAssignExpr}{ArmadilloConnection}(conn, symbol, expr, async = TRUE)
}
\arguments{
\item{conn}{\code{\link{ArmadilloConnection-class}} object.}

\item{symbol}{Name of the R symbol.}

\item{expr}{A R expression with allowed assign functions calls.}

\item{async}{Whether the result of the call should be retrieved
asynchronously. When TRUE (default) the calls are parallelized over the
connections, when the connection supports that feature, with an extra
overhead of requests.}
}
\value{
A \code{\link{ArmadilloResult-class}} object.
}
\description{
Assign a result of the execution of an expression in the DataSHIELD R
session.
}
