% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloConnection.R
\name{dsAssignResource,ArmadilloConnection-method}
\alias{dsAssignResource,ArmadilloConnection-method}
\title{Assign a resource}
\usage{
\S4method{dsAssignResource}{ArmadilloConnection}(conn, symbol, resource, async = TRUE)
}
\arguments{
\item{conn}{An object that inherits from \code{\link[DSI]{DSConnection-class}}.}

\item{symbol}{Name of the R symbol.}

\item{resource}{Fully qualified name of a resource reference in the data
repository.}

\item{async}{Whether the result of the call should be retrieved
asynchronously.}
}
\value{
A \code{\link{ArmadilloResult-class}} object.
}
\description{
Assign a resource in the DataSHIELD R session.
}
