% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloDriver.R
\name{dsConnect,ArmadilloDriver-method}
\alias{dsConnect,ArmadilloDriver-method}
\title{Connect to an Armadillo DataSHIELD service}
\usage{
\S4method{dsConnect}{ArmadilloDriver}(
  drv,
  name,
  restore = NULL,
  username = "",
  password = "",
  token = "",
  url,
  profile = "default",
  opts = list(),
  ...
)
}
\arguments{
\item{drv}{\code{\link{ArmadilloDriver-class}} class object.}

\item{name}{Name of the connection, which must be unique among all the
DataSHIELD connections.}

\item{restore}{Workspace name to be restored in the newly created DataSHIELD
R session.}

\item{username}{The username to authenticate with.}

\item{password}{The password to authenticate with.}

\item{token}{The ID token to authenticate with.}

\item{url}{URL of the server.}

\item{profile}{the profile to select, default "default"}

\item{opts}{Curl options as described by httr (call httr::httr_options()
for details). Can be provided by "Armadillo.opts" option.}

\item{...}{Unused, needed for compatibility with generic.}
}
\value{
A \code{\link{ArmadilloConnection-class}} object.
}
\description{
Connect to an Armadillo DataSHIELD service, with provided credentials.
}
