% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloConnection.R
\name{dsDisconnect,ArmadilloConnection-method}
\alias{dsDisconnect,ArmadilloConnection-method}
\title{Disconnect from an Armadillo DataSHIELD Service}
\usage{
\S4method{dsDisconnect}{ArmadilloConnection}(conn, save = NULL)
}
\arguments{
\item{conn}{\code{\link{ArmadilloConnection-class}} class
object}

\item{save}{Save the DataSHIELD R session with provided ID (must be a
character string).}
}
\description{
Disconnect from an Armadillo DataSHIELD Service and release all R resources.
If a workspace ID is provided, the DataSHIELD R session will be saved before
being destroyed.
}
\seealso{
\code{\link[DSI]{dsDisconnect}}
}
