% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloResult.R
\name{dsFetch,ArmadilloResult-method}
\alias{dsFetch,ArmadilloResult-method}
\title{Fetch the result}
\usage{
\S4method{dsFetch}{ArmadilloResult}(res)
}
\arguments{
\item{res}{\code{\link{ArmadilloResult-class}} object.}
}
\value{
TRUE if table exists.
}
\description{
Fetch the DataSHIELD operation result.
}
