% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloConnection.R
\name{dsGetInfo,ArmadilloConnection-method}
\alias{dsGetInfo,ArmadilloConnection-method}
\title{Get connection info}
\usage{
\S4method{dsGetInfo}{ArmadilloConnection}(dsObj, ...)
}
\arguments{
\item{dsObj}{\code{\link{ArmadilloConnection-class}} class object}

\item{...}{Unused, needed for compatibility with generic.}
}
\value{
The connection information. This should report the version of
the data repository application (`repo.version`) and its name (`repo.name`),
the database name (`dbname`), username, (`username`), host (`host`), port
(`port`), etc.
It MAY also include any other arguments related to the connection
(e.g., thread id, socket or TCP connection type). It MUST NOT include the
password.
}
\description{
Get information about a connection.
}
