% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloDriver.R
\name{dsGetInfo,ArmadilloDriver-method}
\alias{dsGetInfo,ArmadilloDriver-method}
\title{Get driver info}
\usage{
\S4method{dsGetInfo}{ArmadilloDriver}(dsObj, ...)
}
\arguments{
\item{dsObj}{\code{\link{ArmadilloDriver-class}} class object}

\item{...}{Unused, needed for compatibility with generic.}
}
\value{
The connection information. This returns the version of the
package (`driver.version`). There is no underlying client library.
}
\description{
Get information about a driver.
}
