% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloResult.R
\name{dsGetInfo,ArmadilloResult-method}
\alias{dsGetInfo,ArmadilloResult-method}
\title{Get result info}
\usage{
\S4method{dsGetInfo}{ArmadilloResult}(dsObj, ...)
}
\arguments{
\item{dsObj}{\code{\link{ArmadilloResult-class}} class object}

\item{...}{Unused, needed for compatibility with generic.}
}
\value{
The result information. This should include the R expression
being executed (`expression`) and if the query is complete
(`\{ "status" = "COMPLETED" \}`).
}
\description{
Await completion and get the information about a command
(if still available).
}
