% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloConnection.R
\name{dsHasResource,ArmadilloConnection-method}
\alias{dsHasResource,ArmadilloConnection-method}
\title{Verify resource exists.}
\usage{
\S4method{dsHasResource}{ArmadilloConnection}(conn, resource)
}
\arguments{
\item{conn}{\code{\link{ArmadilloConnection-class}} class object.}

\item{resource}{The identifier of the resource}
}
\value{
TRUE if resource exists.
}
\description{
Verify resource exists.
}
