% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloConnection.R
\name{dsIsAsync,ArmadilloConnection-method}
\alias{dsIsAsync,ArmadilloConnection-method}
\title{Armadillo DataShield Service asynchronous support}
\usage{
\S4method{dsIsAsync}{ArmadilloConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{ArmadilloConnection-class}} class object}
}
\value{
The named list of logicals detailing the asynchronicity support.
}
\description{
List of DataSHIELD operations on which Armadillo DataSHIELD Service supports
asynchronicity.
}
\details{
When a \code{\link[DSI]{DSResult-class}} object is returned on aggregation or
assignment operation, the raw result can be accessed asynchronously,
allowing parallelization of DataSHIELD calls over multpile servers.
The returned named list of logicals will specify if asynchronicity is
supported for:
aggregation operation ('aggregate'),
table assignment operation ('assignTable'),
resource assignment operation ('assignResource')
and expression assignment operation ('assignExpr').
}
