% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloResult.R
\name{dsIsCompleted,ArmadilloResult-method}
\alias{dsIsCompleted,ArmadilloResult-method}
\title{Get whether the operation is completed}
\usage{
\S4method{dsIsCompleted}{ArmadilloResult}(res)
}
\arguments{
\item{res}{\code{\link{ArmadilloResult-class}} object.}
}
\value{
TRUE if operation is completed.
}
\description{
Get whether the result from a previous assignment or aggregation operation
was completed, either with a successful status or a failed one. This call
must not wait for the completion, immediate response is expected. Once the
result is identified as being completed, the raw result the operation can be
got directly.
}
