% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloConnection.R
\name{dsKeepAlive,ArmadilloConnection-method}
\alias{dsKeepAlive,ArmadilloConnection-method}
\title{Keep a connection alive}
\usage{
\S4method{dsKeepAlive}{ArmadilloConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{ArmadilloConnection-class}} class object}
}
\value{
NULL, invisibly
}
\description{
As the DataSHIELD sessions are working in parallel, this function helps at
keeping idle connections alive while others are working. Any communication
failure must be silently processed.
}
