% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloConnection.R
\name{dsListMethods,ArmadilloConnection-method}
\alias{dsListMethods,ArmadilloConnection-method}
\title{List methods}
\usage{
\S4method{dsListMethods}{ArmadilloConnection}(conn, type = "aggregate")
}
\arguments{
\item{conn}{\code{\link{ArmadilloConnection-class}} class object}

\item{type}{Type of the method: "aggregate" (default) or "assign".}
}
\value{
A data.frame with columns: name, type ('aggregate' or 'assign'),
class ('function' or 'script'), value, package, version.
}
\description{
List methods defined in the DataSHIELD configuration.
}
