% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloConnection.R
\name{dsListPackages,ArmadilloConnection-method}
\alias{dsListPackages,ArmadilloConnection-method}
\title{List packages}
\usage{
\S4method{dsListPackages}{ArmadilloConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{ArmadilloConnection-class}} class object}
}
\value{
A data.frame with columns: name, version.
}
\description{
List packages with their versions defined in the DataSHIELD configuration.
}
