% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloConnection.R
\name{dsListResources,ArmadilloConnection-method}
\alias{dsListResources,ArmadilloConnection-method}
\title{List Armadillo DataSHIELD Service resources}
\usage{
\S4method{dsListResources}{ArmadilloConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{ArmadilloConnection-class}} class object}
}
\value{
The fully qualified names of the resources.
}
\description{
List Armadillo DataSHIELD Service resources that may be accessible for
performing DataSHIELD operations.
}
\seealso{
\code{\link[DSI]{dsListResources}}
}
