% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloConnection.R
\name{dsListSymbols,ArmadilloConnection-method}
\alias{dsListSymbols,ArmadilloConnection-method}
\title{List R symbols}
\usage{
\S4method{dsListSymbols}{ArmadilloConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{ArmadilloConnection-class}} class object}
}
\value{
A character vector.
}
\description{
List symbols living in the DataSHIELD R session.
}
