% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloConnection.R
\name{dsListTables,ArmadilloConnection-method}
\alias{dsListTables,ArmadilloConnection-method}
\title{List Armadillo DataSHIELD Service tables}
\usage{
\S4method{dsListTables}{ArmadilloConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{ArmadilloConnection-class}} class object}
}
\value{
The fully qualified names of the tables.
}
\description{
List Armadillo DataSHIELD Service tables that may be accessible for
performing DataSHIELD operations.
}
\seealso{
\code{\link[DSI]{dsListTables}}
}
