% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloConnection.R
\name{dsListWorkspaces,ArmadilloConnection-method}
\alias{dsListWorkspaces,ArmadilloConnection-method}
\title{List workspaces}
\usage{
\S4method{dsListWorkspaces}{ArmadilloConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{ArmadilloConnection-class}} class object}
}
\value{
A data.frame with columns: name, lastAccessDate, size, user.
}
\description{
List workspaces saved in the data repository.
}
