% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSMolgenisArmadillo-package.R
\docType{package}
\name{DSMolgenisArmadillo-package}
\alias{DSMolgenisArmadillo}
\alias{DSMolgenisArmadillo-package}
\title{DSMolgenisArmadillo: 'DataSHIELD' Client for 'MOLGENIS Armadillo'}
\description{
'DataSHIELD' is an infrastructure and series of R packages that enables the remote and 'non-disclosive' analysis of sensitive research data. This package is the 'DataSHIELD' interface implementation to analyze data shared on a 'MOLGENIS Armadillo' server. 'MOLGENIS Armadillo' is a light-weight 'DataSHIELD' server using a file store and an 'RServe' server.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/molgenis/molgenis-r-datashield/}
  \item \url{https://molgenis.github.io/molgenis-r-datashield/}
  \item Report bugs at \url{https://github.com/molgenis/molgenis-r-datashield/issues/}
}

}
\author{
\strong{Maintainer}: Mariska Slofstra \email{m.k.slofstra@umcg.nl} (\href{https://orcid.org/0000-0002-0400-0468}{ORCID})

Authors:
\itemize{
  \item Sido Haakma \email{s.haakma@rug.nl} (\href{https://orcid.org/0000-0003-1278-9144}{ORCID})
  \item Tommy de Boer \email{t.de.boer01@umcg.nl} (\href{https://orcid.org/0000-0002-4492-7225}{ORCID})
  \item Fleur Kelpin \email{fleur@kelpin.nl} (\href{https://orcid.org/0000-0003-1527-5130}{ORCID})
}

Other contributors:
\itemize{
  \item MOLGENIS org \email{molgenis-support@umcg.nl} [copyright holder, funder]
}

}
\keyword{internal}
