% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloConnection.R
\name{dsRmWorkspace,ArmadilloConnection-method}
\alias{dsRmWorkspace,ArmadilloConnection-method}
\title{Remove a workspace}
\usage{
\S4method{dsRmWorkspace}{ArmadilloConnection}(conn, name)
}
\arguments{
\item{conn}{\code{\link{ArmadilloConnection-class}} class object}

\item{name}{Name of the workspace.}
}
\description{
Remove a workspace on the data repository.
}
