% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloConnection.R
\name{dsSaveWorkspace,ArmadilloConnection-method}
\alias{dsSaveWorkspace,ArmadilloConnection-method}
\title{Save workspace}
\usage{
\S4method{dsSaveWorkspace}{ArmadilloConnection}(conn, name)
}
\arguments{
\item{conn}{\code{\link{ArmadilloConnection-class}} class object}

\item{name}{Name of the workspace.}
}
\description{
Save workspace on the data repository.
}
